aider: Added psutil to requirements.txt to resolve ModuleNotFoundError for psutil.
# Aider chat conversation:

USER: I ran this command:

 python test.py

And got this output:

Traceback (most recent call last):
  File "/Users/djh/work/src/github.com/djh00t/module_klingon_serial/test.py", line 1, in <module>
    import klingon_serial
  File "/Users/djh/work/src/github.com/djh00t/module_klingon_serial/klingon_serial/__init__.py", line 2, in <module>
    from .utils import get_mac_address_and_interface, get_debug
  File "/Users/djh/work/src/github.com/djh00t/module_klingon_serial/klingon_serial/utils.py", line 2, in <module>
    import psutil
ModuleNotFoundError: No module named 'psutil'

ASSISTANT: The error you're encountering is because the `psutil` module is not installed in your Python environment. To resolve this, you need to add `psutil` to your `requirements.txt` file so that it can be installed along with the other dependencies.

Here is the diff for the change to `requirements.txt`:

```diff
--- requirements.txt
+++ requirements.txt
@@ ... @@
+psutil
```

After applying this patch, you should run `pip install -r requirements.txt` to install the newly added `psutil` module.