aider: Add support for passing 'Accept' header as a query parameter in Swagger UI.
# Aider chat conversation:

USER: Swagger (when I access it via /docs endpoint) is sending the accept header i provide it with as a URL argument instead of as a request header. How do we fix this?

ASSISTANT: To address the issue with Swagger UI sending the `Accept` header as a URL argument instead of as a request header, we need to modify the `root` endpoint in the `handler.py` file to correctly interpret the `Accept` header when it's provided as a query parameter. This is not the standard behavior for Swagger UI, but we can add support for it in our endpoint.

Here are the diffs for the changes to the `handler.py` file:

```diff
--- openfaas/handler.py
+++ openfaas/handler.py
@@ ... @@
 @app.get("/")
-async def root(accept: Optional[str] = Header(None)):
+async def root(accept: Optional[str] = Header(None), accept_query: Optional[str] = Query(None, alias='Accept')):
     # Root endpoint that generates and returns a unique serial number in the requested format.
     # The Accept header determines the response content type: JSON, plain text, HTML, XML, or XHTML.
     # If the Accept header is not supported, it returns a 406 Not Acceptable with an error message.
     unique_serial = generate_serial().upper()
     data = {"serial": unique_serial}
+    # Check if 'Accept' is provided as a query parameter and override the header value
+    if accept_query:
+        accept = accept_query
     if accept:
         if "application/json" in accept:
             return JSONResponse(content=data)
         elif "text/plain" in accept:
             return PlainTextResponse(unique_serial)
         elif "text/html" in accept:
             html_content = f'<html><body><p>{unique_serial}</p></body></html>'
             return HTMLResponse(content=html_content)
         elif "application/xml" in accept:
             xml_content = f'<root><serial>{unique_serial}</serial></root>'
             return XMLResponse(content=xml_content)
         elif "application/xhtml+xml" in accept:
             xhtml_content = f'<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Serial Number</title></head><body><p>{unique_serial}</p></body></html>'
             return HTMLResponse(content=xhtml_content)
         elif "application/yaml" in accept:
             yaml_content = yaml.dump(data)
             return YAMLResponse(content=yaml_content)
     raise HTTPException(status_code=406, detail="Unsupported Accept header")
```

This change adds a new query parameter `accept_query` to the `root` endpoint, which is an alias for `Accept`. If `accept_query` is provided, it will override the `accept` header value. This allows Swagger UI to pass the `Accept` header as a query parameter, and the server will handle it as if it were passed as a header.